+++ outputfilename.txt
@@ -1,0 +1,30 @@
+     x  = 6
+     y  = 7
+displacement_point_force_P_wave_output
+        -0	-3.23714e-05	-6.47428e-05	-9.71142e-05	-0.000129486	-0.000161857	-0.000194228	-0.0002266	-0.000258971	
+displacement_single_couple_P_wave_output
+        -0	-8.44427e-08	-1.68885e-07	-2.53328e-07	-3.37771e-07	-4.22213e-07	-5.06656e-07	-5.91099e-07	-6.75542e-07	
+displacement_double_couple_P_wave_output
+        -0	-1.68885e-07	-3.37771e-07	-5.06656e-07	-6.75542e-07	-8.44427e-07	-1.01331e-06	-1.1822e-06	-1.35108e-06	
+displacement_force_dipole_P_wave_output
+         0	1.66692e-06	3.33385e-06	5.00077e-06	6.66769e-06	8.33462e-06	1.00015e-05	1.16685e-05	1.33354e-05	
+     x  = 6
+     y  = 7
+displacement_point_force_S_wave_output
+         0	         0	         0	         0	         0	         0	         0	         0	         0	
+displacement_single_couple_SV_wave_output
+        -0	-1.30581e-06	-2.61161e-06	-3.91742e-06	-5.22322e-06	-6.52903e-06	-7.83484e-06	-9.14064e-06	-1.04464e-05	
+displacement_double_couple_SV_wave_output
+        -0	-2.61161e-06	-5.22322e-06	-7.83484e-06	-1.04464e-05	-1.30581e-05	-1.56697e-05	-1.82813e-05	-2.08929e-05	
+displacement_force_dipole_SV_wave_output
+         0	0.000603054	0.00120611	0.00180916	0.00241221	0.00301527	0.00361832	0.00422138	0.00482443	
+     x  = 6
+     y  = 7
+displacement_point_force_S_wave_output
+         0	         0	         0	         0	         0	         0	         0	         0	         0	
+displacement_single_couple_SH_wave_output
+         0	3.31437e-08	6.62874e-08	9.94312e-08	1.32575e-07	1.65719e-07	1.98862e-07	2.32006e-07	2.6515e-07	
+displacement_double_couple_SH_wave_output
+        -0	-2.57645e-05	-5.15291e-05	-7.72936e-05	-0.000103058	-0.000128823	-0.000154587	-0.000180352	-0.000206116	
+displacement_force_dipole_SH_wave_output
+        -0	-1.30853e-06	-2.61707e-06	-3.9256e-06	-5.23413e-06	-6.54267e-06	-7.8512e-06	-9.15973e-06	-1.04683e-05
