+++ prototypes.h
@@ -51,103 +51,3 @@ double mesh_gen (double x, double y, double dx, double dy);
 
 double cart_2_sph (double *x, double *y);
 
-// Short description: This function generates P-, SH- and SV-wave radiation patterns for 
-// single couple force, double couple, force dipole and point forces using the values of 
-// theta and phi
-// -------------------------------------------------------------------------------------
-
-double radp_Pw_si_cpl (double theta, double phi);
-
-double radp_SHw_si_cpl (double theta, double phi);
-
-double radp_SVw_si_cpl (double theta, double phi);
-
-double radp_Pw_do_cpl (double theta, double phi);
-
-double radp_SHw_do_cpl (double theta, double phi);
-
-double radp_SVw_do_cpl (double theta, double phi);
-
-double radp_Pw_fo_dipo (double theta, double phi);
-
-double radp_SHw_fo_dipo (double theta, double phi);
-
-double radp_SVw_fo_dipo (double theta, double phi);
-
-double radp_Pw_pt_fo (double theta, double phi);
-
-double radp_SHw_pt_fo (double theta, double phi);
-
-double radp_SVw_pt_fo (double theta, double phi);
-
-// Short description: This function calculates the P-, SH- and SH-wave Displacements for 
-// single force, double couple, force dipole and point forces using the values of theta,
-// phi, distance (R), moment (C1), S-wave velocity (beta), density (rho), the input waveform
-// (h) and its derivative, and the time series (time) and its derivative.
-// --------------------------------------------------------------------------------------------
-
-int displ_pt_fo_Pw (
-    double theta, double R, double moment, double alpha,
-    double rho, double *h, double *t, double *displ_pt_fo_Pw_o, int len);
-
-double displ_pt_fo_Sw (
-    double theta, double phi, double R, double moment, double beta,
-    double rho, double *h, double *t, double *displ_pt_fo_Sw_o, int len);
-
-double displ_si_fo_Pw (
-    double theta, double phi, double R, double moment, double alpha,
-    double rho, double *h_der, double *t_der, double *displ_si_fo_Pw_o, int len);
-
-double displ_si_fo_SHw (
-    double theta, double phi, double R, double moment, double alpha,
-    double rho, double *h_der, double *t_der, double *displ_si_fo_SHw_o, int len);
-
-double displ_si_fo_SVw (
-    double theta, double phi, double R, double moment, double beta,
-    double rho, double *h_der, double *t_der, double *displ_si_fo_SVw_o, int len);
-
-double displ_do_cpl_Pw (
-    double theta, double phi, double R, double moment, double alpha,
-    double rho, double *h_der, double *t_der, double *displ_do_cpl_Pw_o, int len);
-
-double displ_do_cpl_SHw (
-    double theta, double phi, double R, double moment, double alpha,
-    double rho, double *h_der, double *t_der, double *displ_do_cpl_SHw_o, int len);
-
-double displ_do_cpl_SVw (
-    double theta, double phi, double R, double moment, double alpha,
-    double rho, double *h_der, double *t_der, double *displ_do_cpl_SVw_o, int len);
-
-double displ_fo_dipo_Pw (
-    double theta, double phi, double R, double moment, double alpha,
-    double rho, double *h, double *t,  double *displ_fo_dipo_Pw_o, int len);
-
-double displ_fo_dipo_SHw (
-    double theta, double phi, double R, double moment, double alpha,
-    double rho, double *h, double *t,  double *displ_fo_dipo_SHw_o, int len);
-
-double displ_fo_dipo_SVw (
-    double theta, double phi, double R, double moment, double alpha,
-    double rho, double *h, double *t,  double *displ_fo_dipo_SVw_o, int len);	
-		
-// Short description: This function writes the all the P-, SH- and SV-wave displacements,
-// input waveform and its derivative and its location (x,y) into a file
-// --------------------------------------------------------------------------------------		
-
-int wr_Pw_2_file (
-    double *displ_pt_fo_Pw_o, double *displ_si_cpl_Pw_o,
-    double *displ_do_cpl_Pw_o, double *displ_fo_dipo_Pw_o,
-    double *h, double *h_der, double *t, double *t_der, double xx,
-    double yy, string outputfilename, int len);
-
-int wr_SVw_2_file (
-    double *displ_pt_fo_Sw_o, double *displ_si_cpl_SVw_o,
-    double *displ_do_cpl_SVw_o, double *displ_fo_dipo_SVw_o,
-    double *h, double *h_der, double *t, double *t_der, double xx,
-    double yy, string outputfilename, int len);
-
-int wr_SHw_2_file (
-    double *displ_pt_fo_Sw_o, double *displ_si_cpl_SHw_o,
-    double *displ_do_cpl_SHw_o, double *displ_fo_dipo_SHw_o,
-    double *h, double *h_der, double *t, double *t_der, double xy,
-    double yy, string outputfilename, int len);
