+++ radiation.cpp
@@ -1,0 +1,253 @@
+/**
+ * @short   Radiation pattern equations
+ * @file   radiation.cpp
+ * @author Eric Jambo
+ *
+ * This file contains the prototypes and a short description of all the functions used
+ * to calculate Radiation pattern.
+ */
+
+#include <iostream>
+#include <string>
+#include <fstream>
+#include <sstream>
+#include <stdlib.h>
+#include <cstdlib>
+#include <cmath>
+#include "radiation.h"
+
+using namespace std;
+
+// defining the value of PI.
+
+#define PI 3.14159265
+/**
+ * Author:            Eric Jambo
+ *
+ * Short description: This function generates P-wave radiation pattern for single couple
+ *                    force using the values of theta and phi
+ *
+ * Return             0 on sucess
+ *
+ */
+
+double radp_Pw_si_cpl (double theta, double phi)
+{
+
+    cout << "Running: radiation_pattern_P_wave_single_couple_force\n";
+    cout << endl;
+
+    return 0;
+}
+
+/**
+ * Author:            Eric Jambo
+ *
+ * Short description: This function generates SH-wave radiation pattern for single couple
+ *                    force using the values of theta and phi
+ *
+ * Return             0 on sucess
+ *
+ */
+
+double radp_SHw_si_cpl (double theta, double phi)
+{
+
+    cout << "radiation_pattern_SH_wave_single_couple_force\n";
+    cout << endl;
+
+    return 0;
+}
+
+/**
+ * Author:            Eric Jambo
+ *
+ * Short description: This function generates SV-wave radiation pattern for single couple
+ *                    force using the values of theta and phi
+ *
+ * Return             0 on sucess
+ *
+ */
+
+double radp_SVw_si_cpl (double theta, double phi)
+{
+
+    cout << "radiation_pattern_SV_wave_single_couple_force\n";
+    cout << endl;
+
+    return 0;
+}
+
+/**
+ * Author:            Eric Jambo
+ *
+ * Short description: This function generates P-wave radiation pattern for double couple
+ *                    force using the values of theta and phi
+ *
+ * Return             0 on sucess
+ *
+ */
+
+double radp_Pw_do_cpl (double theta, double phi)
+{
+
+    cout << "Running: radiation_pattern_P_wave_double_couple_force\n";
+    cout << endl;
+
+    return 0;
+}
+
+/**
+ * Author:            Eric Jambo
+ *
+ * Short description: This function generates SH-wave radiation pattern for double couple
+ *                    force using the values of theta and phi
+ *
+ * Return             0 on sucess
+ *
+ */
+
+double radp_SHw_do_cpl (double theta, double phi)
+{
+
+    cout << "radiation_pattern_SH_wave_double_couple_force\n";
+    cout << endl;
+
+    return 0;
+}
+
+
+/**
+ * Author:            Eric Jambo
+ *
+ * Short description: This function generates SV-wave radiation pattern for double couple
+ *                    force using the values of theta and phi
+ *
+ * Return             0 on sucess
+ *
+ */
+
+double radp_SVw_do_cpl (double theta, double phi)
+
+{
+
+    cout << "Running: radiation_pattern_SV_wave_double_couple_force\n";
+    cout << endl;
+
+    return 0;
+}
+
+/**
+ * Author:            Eric Jambo
+ *
+ * Short description: This function generates P-wave radiation pattern for force dipole
+ *                    using the values of theta and phi
+ *
+ * Return             0 on sucess
+ *
+ */
+
+double radp_Pw_fo_dipo (double theta, double phi)
+{
+
+    cout << "Running: radiation_pattern_P_wave_force_dipole\n";
+    cout << endl;
+
+    return 0;
+}
+
+/**
+ * Author:            Eric Jambo
+ *
+ * Short description: This function generates SH-wave radiation pattern for force dipole
+ *                    using the values of theta and phi
+ *
+ * Return             0 on sucess
+ *
+ */
+
+double radp_SHw_fo_dipo (double theta, double phi)
+
+{
+
+    cout << "radiation_pattern_SH_wave_force_dipole\n";
+    cout << endl;
+
+    return 0;
+}
+
+/**
+ * Author:            Eric Jambo
+ *
+ * Short description: This function generates SV-wave radiation pattern for force dipole
+ *                    using the values of theta and phi
+ *
+ * Return             0 on sucess
+ *
+ */
+
+double radp_SVw_fo_dipo (double theta, double phi)
+{
+
+    cout << "radiation_pattern_SV_wave_force_dipole\n";
+    cout << endl;
+
+    return 0;
+}
+
+/**
+ * Author:            Eric Jambo
+ *
+ * Short description: This function generates P-wave radiation pattern for point force
+ *                    using the values of theta and phi
+ *
+ * Return             0 on sucess
+ *
+ */
+
+double radp_Pw_pt_fo (double theta, double phi)
+{
+
+    cout << "radiation_pattern_P_wave_point_force\n";
+    cout << endl;
+
+    return 0;
+}
+
+/**
+ * Author:            Eric Jambo
+ *
+ * Short description: This function generates SH-wave radiation pattern for point force
+ *                    using the values of theta and phi
+ *
+ * Return             0 on sucess
+ *
+ */
+
+double radp_SHw_pt_fo (double theta, double phi)
+{
+
+    cout << "radiation_pattern_SH_wave_point_force\n";
+    cout << endl;
+
+    return 0;
+}
+
+/**
+ * Author:            Eric Jambo
+ *
+ * Short description: This function generates SV-wave radiation pattern for point force
+ *                    using the values of theta and phi
+ *
+ * Return             0 on sucess
+ *
+ */
+
+double radp_SVw_pt_fo (double theta, double phi)
+{
+
+    cout << "radiation_pattern_SV_wave_point_force\n";
+    cout << endl;
+
+    return 0;
+}
